Change the 'clean' option to avoid deleting the container folder
Now if you run the script with the "--clean" option it will remove only the generated assets and density folders if they are empty.
This way we avoid deleting images that were added manually.
The density folder and the output folder will be removed only if both are empty after removing the generated assets.

This commit also adds support for a "--force-clean" option.
This option will remove the density generated folders, even if there are files not removed in the clean process (the folder isn't empty).

In both cases the output directory won't be removed, unless it's empty after the clean process.

Also this commit bumps the version to "1.0.1"