Add multithreading support
This commit adds multithreading support for the processing of new and existing files.
The multiprocessing was made ussing pools of workers to process every asset in a worker, with some workers running un parallel.

The number of workers can be changed with a new command line option "--threads" or "-T".
By default the number of threads is "cpu_cores + 1".